Changed the path of the config.properties file
Used a URL object to get the path of the config.properties file. Moved the config.properties file to the resources folder under src so it would be able to be found by .getResource. Changed the if statement to now check if the URL returns a null path instead. Cleaned up the test.